Move LabyrinthGenerator to generator.py
